Add two ways to get resource from the storage: shallow or not.
Shallow means that the, if it's collection resource, it will not include its child VEVENTs in the ICS data.
This is used throughout the palces where the children dont matter.

Not shallow means that the child VEVENTs will be included in the returned ICS content data for collection resources.
This is used then sending a GET request to fetch a specific resource and expecting its full ICS data in response.