Add REPORT unit tests.
Move the tests for REPORT covered in the integration tests to the unit test of REPORT.
The tests are not really isolated, as they are still touching the storage code, but that's fair for now.
Now the REPORT tests in the integration tests are smaller and cover just a happy path, working
more like a smoke test to check all the funnel is working. The different REPORT scenarios are now covered
in the report unit tests.

NOTE: This commit introduces a new REPORT test scenario, which is failing now, as the current code base has a bug.
      The bug will be fixed in PR #7 (2359fcc1fed71f0383e9013ea6bc1459f887d35f)